Tests/LibWeb/Text/expected/DOM/Range-containing-CDATASection.txt

range start offset: 0, end offset: 3
